requirements.txt
django==6.0.1
djangorestframework==3.16.1

psycopg[binary]==3.3.2

python-dotenv==1.1.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0

pytest==9.0.2
pytest-django==4.11.1

drf-yasg==1.21.11
django-cors-headers==4.9.0
whitenoise==6.11.0
pillow==12.1.0
